### 2026-02-25 #754 @maelle
saperlipopette: a risk-free playground for learning more Git
Submitting Author Name: Maëlle Salmon
Submitting Author Github Handle: @maelle
Repository:  https://github.com/ropensci-training/saperlipopette
Version submitted: 0.1.1
Submission type: Standard

Version accepted: TBD

---

```
Package: saperlipopette
Title: Create Example Git Messes
Version: 0.1.0.9000
    person("Maëlle", "Salmon", , "msmaellesalmon@gmail.com", role = c("cre", "aut"),
    person("Jim", "Gardner", role = "ctb"),
    person("Yanina Bellini", "Saibene", role = "aut",
  )
Description: Holds functions creating Git messes, that users would then
    solve, to follow <https://ohshitgit.com/>.
    https://github.com/ropensci-training/saperlipopette
    brio,
    cli,
    fs,
    gert,
    parsedate,
    roxygen2,
    usethis,
    withr
    quarto,
    quarto
Config/Needs/website: rmarkdown, quarto
Config/potools/style: explicit
RoxygenNote: 7.3.3.9000
Config/ropensci/maintainer: staff

```

	- [x] rOpenSci internal tools

This package is used to teach Git during the technical training of rOpenSci champions.

The target audience of saperlipopette consists of professionals developing software who usually work with R and know basic Git commands such as add, commit, push, and pull, and who understand a bit about branches, but are not yet fluent, and therefore underuse Git’s palette of features. The package is suitable for instructors to use as a safe practice environment during workshops when live-coding and running hands-on exercises, as well for for self-guided learners.

Not to our knowledge. There are online platforms for learning Git, but with saperlipopette you get exercises in your local environment.

N/A

N/A

We intend to submit a paper about this package to JOSE. https://openjournals.readthedocs.io/en/jose/submitting.html (the process is expedited for packages that have been reviewed for rOpenSci https://openjournals.readthedocs.io/en/jose/editing.html?highlight=ropensci#finding-reviewers, like at JOSS)

## AI usage

Some notes on LLM usage! I resumed work on saperlipopette after I got a Claude Code subscription and after we started drafting the AI policy, so I was particularly mindful of my LLM usage.

I used Claude to generate the latest exercises, based on my instructions on their content and on the infrastructure of existing exercises. I reviewed and tested the exercises. The first PR created in this way also added a checklist to the [`CONTRIBUTING.md` file](https://github.com/ropensci-training/saperlipopette/blob/main/CONTRIBUTING.md). Those commits have Claude as a co-author:

```
git log -i --grep="co-authored-by: claude"
commit a35f825b9c77ff82d064e090665e602c9721cf2d
Author: Maëlle Salmon <maelle.salmon@yahoo.se>
Date:   Fri Feb 13 10:37:09 2026 +0100

    feat: blame exercise (#40)

    Co-authored-by: Claude Opus 4.6

commit a780263dae95ba2ae90d8ed424fd9fe521e02a4f
Author: Maëlle Salmon <maelle.salmon@yahoo.se>
Date:   Fri Feb 13 10:14:41 2026 +0100

    feat: git log exercises (#39)

    Co-authored-by: Claude Opus 4.6

commit 380920fabaf4ceba81bf02b65c29f5b761b71158
Author: Maëlle Salmon <maelle.salmon@yahoo.se>
Date:   Fri Feb 13 09:43:13 2026 +0100

    feat: rev-parse exercise (#38)

    Co-authored-by: Claude Opus 4.6

commit 4a997953d4a71fa4853cb5579be7cd758f2e3d98
Author: Maëlle Salmon <maelle.salmon@yahoo.se>
Date:   Fri Feb 13 09:01:42 2026 +0100

    feat: new worktree exercise (#37)

    Co-authored-by: Claude Opus 4.6
```

I also used Claude for light editing of the paper.md (looking for repetitions) but that paper.md was mostly created and reviewed by me, Yanina Bellini Saibene, Steffi LaZerte.

The paper for JOSE is rendered as a vignette, the vignette is updated by running `tools/update-paper-qmd.R` which is an utility script written by Claude.
