### 2026-02-26 9f0ea40
chore: add @steffilazerte to authors list
Also normalizes `DESCRIPTION`

Fix #47
