+++ DESCRIPTION
@@ -6,7 +6,9 @@ Authors@R: c(
            comment = c(ORCID = "0000-0002-2815-0399")),
     person("Jim", "Gardner", role = "ctb"),
     person("Yanina Bellini", "Saibene", role = "aut",
-           comment = c(ORCID = "0000-0002-4522-7466"))
+           comment = c(ORCID = "0000-0002-4522-7466")),
+    person("Steffi", "LaZerte", role = "rev",
+           comment = c("Steffi reviewed the 'saperlipopette, a risk-free playground for learning more Git' article.", ORCID = "0000-0002-7690-8360"))
   )
 Description: Holds functions creating Git messes, that users would then
     solve, to follow <https://ohshitgit.com/>.
@@ -34,8 +36,8 @@ VignetteBuilder:
     quarto
 Config/Needs/website: rmarkdown, quarto
 Config/potools/style: explicit
+Config/ropensci/maintainer: staff
 Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
 RoxygenNote: 7.3.3.9000
-Config/ropensci/maintainer: staff
