### 2026-02-26 d91cd32
Deploying to gh-pages from @ ropensci-training/saperlipopette@9f0ea401ff9cb6e24ec4456274b0306a4a5ee92a 🚀
