### 2026-03-05 f6add35
docs: remove useless badge
