### 2026-03-05 9d94af9
test: fix code for interactive testing within RStudio/Positron
