+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: saperlipopette
 Title: Create Example Git Messes
-Version: 0.1.0.9000
+Version: 0.1.1
 Authors@R: c(
     person("Maëlle", "Salmon", , "msmaellesalmon@gmail.com", role = c("cre", "aut"),
            comment = c(ORCID = "0000-0002-2815-0399")),
