+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   articles/paper: paper.html
   saperlipopette: saperlipopette.html
-last_built: 2026-03-05T13:30Z
+last_built: 2026-03-05T13:36Z
 urls:
   reference: https://docs.ropensci.org/saperlipopette/reference
   article: https://docs.ropensci.org/saperlipopette/articles
