+++ NEWS.md
@@ -1,3 +1,5 @@
+# saperlipopette (development version)
+
 # saperlipopette 0.1.1
 
 * Fix tests for interactive testing within RStudio or Positron (thanks `@robitalec`).
