### 2026-03-05 a01d1fe
Deploying to gh-pages from @ ropensci-training/saperlipopette@dc4ec9e81fb7f64a4e46502eb5abec89777dc463 🚀
