+++ authors.md
@@ -21,13 +21,12 @@ Source:
 [`DESCRIPTION`](https://github.com/ropensci-training/saperlipopette/blob/main/DESCRIPTION)
 
 Salmon M, Saibene Y (2026). *saperlipopette: Create Example Git Messes*.
-R package version 0.1.0.9000,
-<https://docs.ropensci.org/saperlipopette/>.
+R package version 0.1.1, <https://docs.ropensci.org/saperlipopette/>.
 
     @Manual{,
       title = {saperlipopette: Create Example Git Messes},
       author = {Maëlle Salmon and Yanina Bellini Saibene},
       year = {2026},
-      note = {R package version 0.1.0.9000},
+      note = {R package version 0.1.1},
       url = {https://docs.ropensci.org/saperlipopette/},
     }
