+++ news/index.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## saperlipopette (development version)
+
 ## saperlipopette 0.1.1
 
 - Fix tests for interactive testing within RStudio or Positron (thanks
