### 2026-03-05 #754 @maelle
Thank you @robitalec!!

> I'm seeing two different license related files: [LICENSE](https://github.com/ropensci-training/saperlipopette/blob/9f0ea401ff9cb6e24ec4456274b0306a4a5ee92a/LICENSE) and [LICENSE.md](https://github.com/ropensci-training/saperlipopette/blob/9f0ea401ff9cb6e24ec4456274b0306a4a5ee92a/LICENSE.md).

Yes, one is for R (`LICENSE`) the other for GitHub-readability (`LICENSE.md`). That's what `usethis::use_mit_license()` creates. See also https://r-pkgs.org/license.html#key-files

> We could use a bit more information - right now we see "Please open an issue for any larger change (new exercise for instance).", but what about recommendations for eg. filing bugs, asking for help, contributing smaller changes?

Thank you! I added a bit in a [commit](https://github.com/ropensci-training/saperlipopette/commit/6ae9f802b8fa2587d27178457a1654d7ee79316f) -- do you think that is enough or too little information? It's the first time I start a contributing guide from scratch as opposed to a standard template, maybe that's too minimalist. 😅

> I edited the template to remove these two lines, they didn't seem to correspond to anything relevant:

I think this might be related to recent work by @mpadge to accomodate for packages contained in a subdirectory?

>     When we are ready to look for reviewers, could you edit the template to include the version number under "Version submitted"?

I've now updated this field but can increase the version again if I do more changes before you start looking for reviewers.

> There are two [code coverage badges](https://github.com/ropensci-training/saperlipopette/blob/9f0ea401ff9cb6e24ec4456274b0306a4a5ee92a/README.Rmd#L20-L21) in the README

Oops, [updated](https://github.com/ropensci-training/saperlipopette/commit/f6add350323c782258a2251474fb9b108821072a) now!

> The tool that Claude wrote for keeping the vignette updated based on the paper in inst/. Is it possible you could just use the [Includes](https://quarto.org/docs/authoring/includes.html) syntax instead? This works when I check locally with pkgdown::build_site and a basic render. But I'm not sure if it would work/be accepted everywhere (eg. CRAN).

If I remember correctly that's what I first tried doing but the script does a bit more: changing header levels, copying the picture. Do you think that makes sense?

> I see failing tests locally related to the testthat snapshots of gert::git_log hashes. These don't appear on the tests run through GitHub Actions, so I'm trying to figure them out. I've made sure my packages (gert, testthat) and version of libgit2 are updated. I see 1) mostly differences in Git hashes (eg. test-blame and test-committed-to-main) but also 2) some differences in test-create-all that might be due to snapshots not having RStudio Projects files, whereas my local tests generates them? Let me know what you think might be going on.

Oh yes you are right! I never run `devtools::test()` from my console so hadn't noticed. Once I did, I was able to reproduce the test failures. I have [updated the code](https://github.com/ropensci-training/saperlipopette/commit/9d94af9994a51e351cdd484353b74195b6e9739e) so that it never creates the RStudio project files when testing.

> Tests have a very similar structure for all the exercises. Tests check for 1) the filepath returned from the example function matches the expected filepath name (as in, does dir_create create a directory named as expected) and 2) the snapshot of git commit hashes returned from gert::git_log match the snapshot.

Thank you for experimenting! I prefer seeing explicitly what's done in each test file even if that's repetitive so would prefer not to change but am happy to hear further reasons. :wink:
