### 2026-03-09 493a4ba
docs: add rOpenSci software peer-review badge
