+++ README.md
@@ -8,6 +8,8 @@
 [![R-CMD-check](https://github.com/ropensci-training/saperlipopette/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ropensci-training/saperlipopette/actions/workflows/R-CMD-check.yaml)
 [![Codecov test
 coverage](https://codecov.io/gh/ropensci-training/saperlipopette/graph/badge.svg)](https://app.codecov.io/gh/ropensci-training/saperlipopette)
+[![Status at rOpenSci Software Peer
+Review](https://badges.ropensci.org/754_status.svg)](https://github.com/ropensci/software-review/issues/754)
 <!-- badges: end -->
 
 The goal of saperlipopette is to hold functions creating Git exercises,
@@ -58,10 +60,10 @@ exclamation](https://en.wiktionary.org/wiki/saperlipopette). You can say
 library("saperlipopette")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/Rtmp1FQ4M1/filee8d654dc7ab6/one-small-change!
+#> ℹ Follow along in /tmp/RtmpQZY9FU/file426d67746edf/one-small-change!
 # what's in path
 fs::dir_tree(path)
-#> /tmp/Rtmp1FQ4M1/filee8d654dc7ab6/one-small-change
+#> /tmp/RtmpQZY9FU/file426d67746edf/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
@@ -113,7 +115,7 @@ building documentation.
 ``` r
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/Rtmp1FQ4M1/filee8d6adda1e6/one-small-change!
+#> ℹ Follow along in /tmp/RtmpQZY9FU/file426d250516e2/one-small-change!
 gert::git_log(repo = path)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -122,7 +124,7 @@ gert::git_log(repo = path)
 #> 2 e227ecc55e421f70b6e30602e6a2ee… Jane … 2023-12-15 16:25:00     2 FALSE "First…
 parent_path2 <- withr::local_tempdir()
 path2 <- exo_one_small_change(parent_path2)
-#> ℹ Follow along in /tmp/Rtmp1FQ4M1/filee8d66565386e/one-small-change!
+#> ℹ Follow along in /tmp/RtmpQZY9FU/file426d31c55757/one-small-change!
 gert::git_log(repo = path2)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -141,10 +143,10 @@ library("saperlipopette")
 withr::local_language("fr")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ L'exercice attend dans /tmp/Rtmp1FQ4M1/filee8d66be603fb/one-small-change !
+#> ℹ L'exercice attend dans /tmp/RtmpQZY9FU/file426d1870e821/one-small-change !
 # what's in path
 fs::dir_tree(path)
-#> /tmp/Rtmp1FQ4M1/filee8d66be603fb/one-small-change
+#> /tmp/RtmpQZY9FU/file426d1870e821/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
