### 2026-03-12 24d43c2
fix: add the missing add step
