### 2026-03-19 c37b8b3
fix: escape curly braces
