+++ inst/exo_time_machine-Rprofile.en.R
@@ -14,6 +14,6 @@ cli::cli_alert_info("For more help use {.run tip()}")
 
 tip <- function() {
   cli::cli_li("{.code git reflog}")
-  cli::cli_li('{.code git reset HEAD@{1}}')
+  cli::cli_li('{.code git reset HEAD@{{1}}}')
   cli::cli_li('Examine Git history.')
 }
