+++ inst/exo_time_machine-Rprofile.fr.R
@@ -14,6 +14,6 @@ cli::cli_alert_info("Pour plus d'aide, {.run tip()}")
 
 tip <- function() {
   cli::cli_li("{.code git reflog}")
-  cli::cli_li('{.code git reset HEAD@{1}}')
+  cli::cli_li('{.code git reset HEAD@{{1}}}')
   cli::cli_li('Examiner l\'historique Git.')
 }
