### 2026-04-01 #754 @jcoliver
---

Thank you @maelle, @yabellini, and @steffilazerte for this package. I think it
provides a nice teaching tool to create common challenging situations in Git
and how to resolve those challenges. The package works as expected, and the
majority of my suggestions address opportunities for documentation to make it
easier to use.

##### High-level package documentation

This is a useful package when paired with published lessons; however, it took
me (as a naive user) a while to get to that important point. (It was not made
clear to me until the "Teaching with saperlipopette" section of the second
vignette). There is a brief mention of this early on in the README; that is,
lines 24-25 of README.Rmd hint at the use, but additional information would be
useful. Something along the lines of

> saperlipopette is designed to be used when teaching intermediate Git by
creating commonly-encountered version control challenges and providing guidance
to resolving the challenges.

In addition to this clarification, I think an expanded vignette with use cases
is warranted. To be extremely pedantic, I see this point in the Packaging
Guide (1.7.1 point #4):

> The package should contain at least one HTML vignette providing a substantial
coverage of package functions, illustrating realistic use cases and how
functions are intended to interact.

Maybe take a look at [https://teachdatascience.com/datapackage/](https://teachdatascience.com/datapackage/)
as an example of a more descriptive vignette on the use of a package in
teaching scenarios.

##### Smaller documentation suggestions

In the package vignette (saperlipopette.qmd), line 42 is:

> At this stage, the user would open the newly created R project and launch an
R session,

But I was surprised because immediately after running
`exo_one_small_change(parent_path)`, a new R project was created *for* me and
another instance of RStudio opened with said project (one_small_change.Rproj).
That is, the user did not have to do anything. I found this behavior startling,
as the documentation indicates that I would have to take some action to open
the project. I think a note about the coming behavior should be added *before*
the code block, so the user has a better sense of what is about to happen. This
could be done by moving lines 42-44 before the code block (and revising the
wording to indicate the things that happen automatically).

The remaining text in this section:

> In practice here the user would change a file, then Git add it, then run `git commit --amend --no-edit`.
The user would examine the [Git history](https://git-scm.com/book/en/v2/Git-Basics-Viewing-the-Commit-History) before and after this.

could use additional details, such as which RStudio instance things should be
happening (because a user will have at least two R/RStudio sessions running at
this point), how to add a change to the registry (will the RStudio Git
interface work or does it need to be done via CLI?), and where the commit
command needs to be run (some users may try to run it in a R console). A nod to
*how* a user would examine the git history (`git log`?) rather than only
referencing an external resource would be a good way to end this section.

Related, all functions that start a new R (and RStudio) session should
explicitly say this behavior (starting a new R session) is expected (probably
best to be done in **Details** sections of function documentation). Yes, I
recognize this would be repetitive, but I think it is useful for new users.

**Specific functions/files**

+ create_all_exercises. I don't know when one would use this function. A use case
(not just an example) would be helpful.
+ The following files include functions that lack documentation for use (even
though most are for internal use only):
  + roxygen2.R
  + utils-fs.R
  + utils-git.R
  + utils-potools.R
  + utils-usethis.R

#### File organization

Not all files are named the same as their functions. Although it is fairly
intuitive, it still adds complexity. Why not pre/post-append filenames so they
align with the function names? e.g. blame.R -> exo_blame.R and create_all.R ->
create_all_exercises.R. This only applies to those files with a single
function; e.g. the approach is not applicable to files such as roxygen2.R.

#### Misc. debris

Consider using `return()`. This is probably just a stylistic difference, but it
does make it a little more explicit. Please ignore if this comment sounds like
something a dinosaur would say.

In the "Follow along..." messages, consider removing the trailing exclamation
point (!). Users may think this needs to be part of the path.

Me, acting *very* naively, identified an opportunity for some defensive
programming: If I try creating all exercises after running any of the other,
individual exercise functions (`exo_*`), it throws an error (appropriately so,
perhaps). For example, I first ran `exo_one_small_change(parent_path)` as in
the example, then ran `create_all_exercises(parent_path)`

```
parent_path <- withr::local_tempdir()
path <- exo_one_small_change(parent_path)
path <- create_all_exercises(parent_path)
```

I received the following error:

```
Error in `map()`:
ℹ In index: 3.
Caused by error in `exo_one_small_change()`:
! Can't create /tmp/Rtmp2PLQsE/file7bd11c387465/one-small-change as it already
exists.
```

Perhaps there could be a catch in `create_all_exercises()` that looks to see if
the temporary directory (and affiliated subdirectories) already exist, and
instead of an Error, it instead Warns the user. I could also see
`create_all_exercises()` itself throwing an Error, with a more descriptive
message. If one is feeling extremely adventurous, another `setdiff()` in
`create_all_exercises()` with the output of something like
`list.dir(parent_path)` (and some fun path parsing), could further reduce the
function names stored in `funs` that are ultimately passed to `purrr::walk()`.
Note I was *not* adventurous enough to test this last suggestion.

##### `devtools::check()`

I saw a message:

> Installed roxygen2 version (7.3.3) doesn't match declared (7.3.3.9000)

This is presumably thrown because the DESCRIPTION includes the RoxygenNote
specifying version 7.3.3.9000. Is current release (7.3.3) not sufficient?

Also, the run of `devtools::check()` failed during the installation step when
trying to install the dependency parsedate. I went ahead and manually installed
parsedate, and the check proceeded without additional problems. It is entirely
possible I forgot to do something before running `devtools::check()`.

There was also this warning:

> WARNING: Added dependency on R >= 4.1.0 because package code uses the
     pipe |> or function shorthand \(...) syntax added in R 4.1.0.

Perhaps this version dependency should be added to the DESCRIPTION? e.g.

And, of course, `devtools::check()` throws this warning which can, IMHO, be
ignored:

>   WARNING
  ‘qpdf’ is needed for checks on size reduction of PDFs

##### `devtools::test()`

40 of 42 tests passed; 2 tests (test-utils-potools.R:2:3,
test-utils-potools.R:13:3) skipped on Linux. Note the two skipped tests are
explicitly skipped on Linux.
