### 2026-04-02 #754 @maelle
Thanks so much @jcoliver!! I'll wait for the second review to perform changes, but I already see how your comments will make saperlipopette much clearer. 🙏
