### 2026-04-02 #754 @robitalec
Thanks for your detailed review @jcoliver. I appreciate your valuable comments especially those relating to bringing the "how to teach/learn with saperlipopette" to the forefront.

Agreed @maelle, we'll wait for the second review before any changes. Thanks!
