### 2026-04-27 #754 @mine-cetinkaya-rundel
---

`saperlipopette` provides hands-on Git exercises for learning and teaching version control.
The package is well-designed, thoroughly documented, and follows rOpenSci best practices. The multilingual support (English, French, Spanish) is particularly impressive and makes this a valuable tool for international audiences.

Overall, the package is well-designed, and the code is of high quality. A consisent API design is followed (e.g., all functions `exo_*` naming convention and take `parent_path` parameter). Additionally, good organization practices are also employed, e.g., utility functions appropriately separated (`utils-git.R`, `utils-fs.R`, `utils-potools.R`)

#### Package documentation

README:

- The README clearly explains the package's purpose as a companion to ohshitgit.com and includes comprehensive installation instructions as well as prerequisites.
- The justification for the package name is nice, though I might suggest also noting that the Wikipedia link includes pronunciation.
- The working example demonstrating the exercise workflow... works! I tried in Positron and had no issues. However, it's likely that a user who needs Git exercises isn't familiar with the withr package. I recommend explaining the code in the first few lines of the example before the example in the README, and also noting that this will start a new Positron session.
- I like the explicit display of multilingual capabilities with a French example as well as links to recommended Git learning resources.
- Badges for R-CMD-check, test coverage, and rOpenSci review status follow good package development practices.

Vignettes:

- Main vignette (`vignettes/saperlipopette.qmd`) provides a comprehensive introduction that is basically the same as the package README, so the same comments apply here. Additionally, the rendered vignette on the package website has "[31m✖[39m", which I believe is a Quarto issue, but it's distracting nonetheless. Is it possible to suppress them?
- The article (`vignettes/articles/paper.qmd`) provides scholarly context on pedagogy and design rationale.

#### Function documentation

- All exported functions are documented with roxygen2. I like the use of `@git` roxygen tags to relevant Git documentation.
- As the other reviewer suggested, expanding documentation for `create_all_exercises()` would be helpful.
- DESCRIPTION specifies `RoxygenNote: 7.3.3.9000` (development version), but the released version 7.3.3 is typically installed. Is there a reason the dev version is needed? Otherwise,c I recommend sticking to the released version.

This is a high-quality package that fills a genuine need in the R and Git education space. The pedagogical approach of providing low-stakes practice environments is sound, and the implementation is clean and well-documented. The multilingual support significantly extends the package's reach and inspires community contributions for further translations.

I think one disconnect is between the level of user who would most benefit from the package (someone new to git) and the level of expertise needed to follow the simplest exercises. This is addressable by adding additional instructions before the first example in the get started vignette (as suggested above).
