### 2026-04-27 aa279a5
docs: clarify use cases of the package
