### 2026-04-27 6cc1429
docs: clarify use cases of the package
