### 2026-04-27 cbf9099
docs: explain what happens automatically
