### 2026-04-27 59bb6b9
docs: explain what happens automatically
