### 2026-04-27 e71ed48
docs: clarify example
