+++ README.Rmd
@@ -68,6 +68,7 @@ If in RStudio or Positron, the code will even open a new R session.
 If not, the user will need to launch an R session,
 where messages would indicate them what to do, 
 and which URL to follow, to find, in this case, the corresponding ohshitgit entry.
+
 ```{r example}
 library("saperlipopette")
 parent_path <- withr::local_tempdir()
@@ -79,8 +80,12 @@ fs::dir_tree(path)
 gert::git_log(repo = path)
 ```
 
-In practice here the user would change a file, then Git add it, then run `git commit --amend --no-edit`.
-The user would examine the [Git history](https://git-scm.com/book/en/v2/Git-Basics-Viewing-the-Commit-History) before and after this.
+In practice here, in the exercise's folder, the user would 
+- change a file, 
+- then Git add it through the terminal or the interface of their choice, 
+- then run `git commit --amend --no-edit` through the terminal or the interface of their choice.
+
+The user would examine the [Git history](https://git-scm.com/book/en/v2/Git-Basics-Viewing-the-Commit-History) before and after this, either through the terminal (`git log`) or the interface of their choice.
 
 ```{r, echo=FALSE}
 source(system.file(
