### 2026-04-27 a4f90b4
docs: clarify example
