+++ vignettes/saperlipopette.qmd
@@ -45,8 +45,12 @@ fs::dir_tree(path)
 gert::git_log(repo = path)
 ```
 
-In practice here the user would change a file, then Git add it, then run `git commit --amend --no-edit`.
-The user would examine the [Git history](https://git-scm.com/book/en/v2/Git-Basics-Viewing-the-Commit-History) before and after this.
+In practice here, in the exercise's folder, the user would 
+- change a file, 
+- then Git add it through the terminal or the interface of their choice, 
+- then run `git commit --amend --no-edit` through the terminal or the interface of their choice.
+
+The user would examine the [Git history](https://git-scm.com/book/en/v2/Git-Basics-Viewing-the-Commit-History) before and after this, either through the terminal (`git log`) or the interface of their choice.
 
 ```{r, echo=FALSE}
 source(system.file(
