### 2026-04-27 82a9656
docs: document Workflow in each manual page
