+++ R/worktree.R
@@ -11,6 +11,7 @@
 #' @git log worktree tag
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
