+++ R/split-changes.R
@@ -16,6 +16,7 @@
 #'
 #' @git add_patch
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
