+++ R/log-deleted-file.R
@@ -9,6 +9,7 @@
 #' @git log_path
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
