+++ R/one-small-change.R
@@ -8,6 +8,7 @@
 #'
 #' @git amend_no_edit
 #' @return The path
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
