+++ R/clean-dir.R
@@ -14,6 +14,7 @@
 #'
 #' @git clean
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
