+++ R/time-machine.R
@@ -8,6 +8,7 @@
 #' @git reset reflog
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
