+++ R/log-deleted-line.R
@@ -10,6 +10,7 @@
 #' @git log_string log_regex
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
