+++ R/committed-to-wrong-branch.R
@@ -8,6 +8,7 @@
 #' @git cherry_pick reset switch checkout
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
