+++ man/exo_reset.Rd
@@ -25,6 +25,18 @@ Useful links:
 \item \url{https://masalmon.eu/2024/06/11/rewrite-git-history/}
 }
 }
+\section{Workflow}{
+
+Running the function will create the exercise as a new folder in \code{parent_path}.
+If called from RStudio or Positron, the function will open a new R session in
+that IDE.
+If not, the user will need to navigate to the path returned by the function,
+and launch an R session from there.
+The new R session will display messages about what challenge to solve,
+and running the \code{tip()} function from that new R session will display
+additional guiance.
+}
+
 \examples{
 \dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 parent_path <- withr::local_tempdir()
