### 2026-04-27 acf53f5
docs: document Workflow in each manual page
