+++ R/reset.R
@@ -15,6 +15,7 @@
 #'
 #' @git reset_mixed
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
