+++ R/rebase-i.R
@@ -16,6 +16,7 @@
 #'
 #' @git rebase_i
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
