+++ R/debug.R
@@ -13,6 +13,7 @@
 #'
 #' @git bisect
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
