+++ R/revert-commit.R
@@ -8,6 +8,7 @@
 #' @git log revert
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
