+++ R/revert-file.R
@@ -8,6 +8,7 @@
 #' @git log restore checkout commit
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
