+++ R/revparse.R
@@ -10,6 +10,7 @@
 #' @git rev_parse log show
 #'
 #' @return The path to the new project
+#' @inheritSection exo_committed_to_main Workflow
 #' @export
 #'
 #' @examplesIf interactive()
