+++ man/exo_blame.Rd
@@ -16,6 +16,18 @@ The path to the new project
 I want to find who added the line \code{x <- x + 1} to the script and when.
 The tool for that is \verb{git blame path/to/file}.
 }
+\section{Workflow}{
+
+Running the function will create the exercise as a new folder in \code{parent_path}.
+If called from RStudio or Positron, the function will open a new R session in
+that IDE.
+If not, the user will need to navigate to the path returned by the function,
+and launch an R session from there.
+The new R session will display messages about what challenge to solve,
+and running the \code{tip()} function from that new R session will display
+additional guiance.
+}
+
 \examples{
 \dontshow{if (interactive()) withAutoprint(\{ # examplesIf}
 parent_path <- withr::local_tempdir()
