### 2026-04-27 f4c7be6
chore: simplify roxygen2 extension stuff, link to refs
