+++ NAMESPACE
@@ -1,6 +1,8 @@
 # Generated by roxygen2: do not edit by hand
 
 S3method(format,rd_section_gitcommands)
+S3method(roxygen2::roxy_tag_parse,roxy_tag_git)
+S3method(roxygen2::roxy_tag_rd,roxy_tag_git)
 export(create_all_exercises)
 export(exo_bisect)
 export(exo_blame)
