### 2026-04-27 f683c72
docs: document fs helpers
