### 2026-04-27 6cba620
docs: document fs helpers
