### 2026-04-27 5f56f85
docs: document and rename Git helpers
