### 2026-04-27 f405a79
docs: document internationalization helpers
