+++ R/utils-potools.R
@@ -1,7 +1,12 @@
+# Helpers for internationalization
+
+# "Explicit" style
+# https://michaelchirico.github.io/potools/articles/developers.html?q=tr_#explicit-style
 tr_ <- function(...) {
   enc2utf8(gettext(paste0(...), domain = "R-saperlipopette"))
 }
 
+# Use the right Rprofile template for the language
 create_r_profile <- function(exo_name) {
   current_language <- tr_("en")
 
