### 2026-04-27 fb2f831
docs: document usethis wrapper
