+++ man/exo_clean_dir.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/clean-dir.R
+% Please edit documentation in R/exo-clean-dir.R
 \name{exo_clean_dir}
 \alias{exo_clean_dir}
 \title{"Hey I'd like to remove these untracked files I created to test stuff!"}
