+++ man/exo_revparse.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/revparse.R
+% Please edit documentation in R/exo-revparse.R
 \name{exo_revparse}
 \alias{exo_revparse}
 \title{"I want to understand ancestry references!"}
