+++ man/exo_one_small_change.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/one-small-change.R
+% Please edit documentation in R/exo-one-small-change.R
 \name{exo_one_small_change}
 \alias{exo_one_small_change}
 \title{"Oh shit, I committed and immediately realized I need to make one small change!"}
