+++ man/exo_log_deleted_line.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/log-deleted-line.R
+% Please edit documentation in R/exo-log-deleted-line.R
 \name{exo_log_deleted_line}
 \alias{exo_log_deleted_line}
 \title{"I want to find which commit deleted a line!"}
