+++ man/exo_log_deleted_file.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/log-deleted-file.R
+% Please edit documentation in R/exo-log-deleted-file.R
 \name{exo_log_deleted_file}
 \alias{exo_log_deleted_file}
 \title{"I want to find which commit deleted a file!"}
