+++ man/exo_committed_to_main.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/committed-to-main.R
+% Please edit documentation in R/exo-committed-to-main.R
 \name{exo_committed_to_main}
 \alias{exo_committed_to_main}
 \title{"Oh shit, I accidentally committed something to main that should have been on a brand new branch!"}
