+++ man/exo_rebase_i.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/rebase-i.R
+% Please edit documentation in R/exo-rebase-i.R
 \name{exo_rebase_i}
 \alias{exo_rebase_i}
 \title{"Hey I'd like to make my commits in a branch look informative and smart!"}
