+++ man/exo_worktree.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/worktree.R
+% Please edit documentation in R/exo-worktree.R
 \name{exo_worktree}
 \alias{exo_worktree}
 \title{"I need to see what the project looked like at a certain version!"}
