+++ man/exo_bisect.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/debug.R
+% Please edit documentation in R/exo-bisect.R
 \name{exo_bisect}
 \alias{exo_bisect}
 \title{"Hey I'd like to find which commit introduced a bug!"}
