+++ man/exo_revert_file.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/revert-file.R
+% Please edit documentation in R/exo-revert-file.R
 \name{exo_revert_file}
 \alias{exo_revert_file}
 \title{"Oh shit, I need to undo my changes to a file!"}
