+++ man/exo_time_machine.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/time-machine.R
+% Please edit documentation in R/exo-time-machine.R
 \name{exo_time_machine}
 \alias{exo_time_machine}
 \title{"Oh shit, I did something terribly wrong, please tell me git has a magic time machine!?!"}
