+++ man/exo_check_editor.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/check-editor.R
+% Please edit documentation in R/exo-check-editor.R
 \name{exo_check_editor}
 \alias{exo_check_editor}
 \title{"Hey, I don't want to get into Vim by mistake!"}
