+++ man/exo_conflict.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/conflict.R
+% Please edit documentation in R/exo-conflict.R
 \name{exo_conflict}
 \alias{exo_conflict}
 \title{"Hey I'd like to see what merge conflicts look like!"}
