+++ man/exo_undo_commit.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/revert-commit.R
+% Please edit documentation in R/exo-undo-commit.R
 \name{exo_undo_commit}
 \alias{exo_undo_commit}
 \title{"Oh shit, I need to undo a commit from like 5 commits ago!"}
