+++ man/exo_latest_message.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/edit-latest-message.R
+% Please edit documentation in R/exo-latest-message.R
 \name{exo_latest_message}
 \alias{exo_latest_message}
 \title{"Oh shit, I need to change the message on my last commit!"}
