+++ man/exo_committed_to_wrong.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/committed-to-wrong-branch.R
+% Please edit documentation in R/exo-committed-to-wrong.R
 \name{exo_committed_to_wrong}
 \alias{exo_committed_to_wrong}
 \title{"Oh shit, I accidentally committed to the wrong branch!"}
