+++ man/exo_split_changes.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/split-changes.R
+% Please edit documentation in R/exo-split-changes.R
 \name{exo_split_changes}
 \alias{exo_split_changes}
 \title{"Hey I'd like to split these changes to the same file into several commits!"}
