+++ man/exo_reset.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/reset.R
+% Please edit documentation in R/exo-reset.R
 \name{exo_reset}
 \alias{exo_reset}
 \title{"Hey I'd like to restart from scratch"}
