+++ man/exo_blame.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/blame.R
+% Please edit documentation in R/exo-blame.R
 \name{exo_blame}
 \alias{exo_blame}
 \title{"I want to find who added a specific line!"}
