+++ R/exo-log-deleted-line.R
@@ -74,7 +74,7 @@ exo_log_deleted_line <- function(parent_path) {
 
   usethis::local_project(original_dir, force = TRUE)
 
-  cli::cli_alert_info(tr_("Follow along in {path}!"))
+  cli::cli_alert_info(tr_("Follow along in {path}"))
 
   path
 }
