+++ R/exo-worktree.R
@@ -82,7 +82,7 @@ exo_worktree <- function(parent_path) {
 
   usethis::local_project(original_dir, force = TRUE)
 
-  cli::cli_alert_info(tr_("Follow along in {path}!"))
+  cli::cli_alert_info(tr_("Follow along in {path}"))
 
   path
 }
