+++ R/exo-log-deleted-file.R
@@ -92,7 +92,7 @@ exo_log_deleted_file <- function(parent_path) {
 
   usethis::local_project(original_dir, force = TRUE)
 
-  cli::cli_alert_info(tr_("Follow along in {path}!"))
+  cli::cli_alert_info(tr_("Follow along in {path}"))
 
   path
 }
