+++ R/exo-committed-to-wrong.R
@@ -59,7 +59,7 @@ exo_committed_to_wrong <- function(parent_path) {
 
   usethis::local_project(original_dir, force = TRUE)
 
-  cli::cli_alert_info(tr_("Follow along in {path}!"))
+  cli::cli_alert_info(tr_("Follow along in {path}"))
 
   path
 }
