+++ man/create_all_exercises.Rd
@@ -17,6 +17,8 @@ But do not open them!
 Having all the exercises as folders within \code{parent_path} makes it possible
 to resolve them one by one, if one wants to practice all that is available
 with the package.
+Run this function only if there are no exercise folder already created in
+\code{parent_path}: you can use a brand-new folder for instance.
 }
 \section{Workflow}{
 
