### 2026-04-27 b62a145
docs: use roxygen2 from cran
