+++ DESCRIPTION
@@ -40,4 +40,4 @@ Config/ropensci/maintainer: staff
 Config/testthat/edition: 3
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.3.3.9000
+RoxygenNote: 7.3.3
