### 2026-04-27 b21b1b3
chore: add minimal R version
