### 2026-04-27 82cf2ba
chore: add minimal R version
