### 2026-04-27 d0d842c
docs: build README
