### 2026-04-27 ac4b2ff
docs: mention recording
