+++ README.md
@@ -55,9 +55,10 @@ You’ll also need
 
 This package is intended to be a companion to <https://ohshitgit.com/>,
 so its name had to honour the exclamation. “saperlipopette” is an
-[old-fashioned French
-exclamation](https://en.wiktionary.org/wiki/saperlipopette). You can say
-“Saperlipopette, Git!”.
+[old-fashioned French exclamation (link including a recording to hear
+the word’s
+pronunciation)](https://en.wiktionary.org/wiki/saperlipopette). You can
+say “Saperlipopette, Git!”.
 
 ## Example
 
@@ -71,10 +72,10 @@ case, the corresponding ohshitgit entry.
 library("saperlipopette")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/Rtmp941JVU/filec7c0f517938/one-small-change
+#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf7ad80e46/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/Rtmp941JVU/filec7c0f517938/one-small-change
+#> /tmp/RtmpSfbavO/filee7bf7ad80e46/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
@@ -128,7 +129,7 @@ building documentation.
 ``` r
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/Rtmp941JVU/filec7c05d23630/one-small-change
+#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf7d1ccd7/one-small-change
 gert::git_log(repo = path)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -137,7 +138,7 @@ gert::git_log(repo = path)
 #> 2 e227ecc55e421f70b6e30602e6a2ee… Jane … 2023-12-15 16:25:00     2 FALSE "First…
 parent_path2 <- withr::local_tempdir()
 path2 <- exo_one_small_change(parent_path2)
-#> ℹ Follow along in /tmp/Rtmp941JVU/filec7c0121eccff/one-small-change
+#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf79c2b66/one-small-change
 gert::git_log(repo = path2)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -156,10 +157,10 @@ library("saperlipopette")
 withr::local_language("fr")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/Rtmp941JVU/filec7c0626d9b81/one-small-change
+#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf36801ecc/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/Rtmp941JVU/filec7c0626d9b81/one-small-change
+#> /tmp/RtmpSfbavO/filee7bf36801ecc/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
