### 2026-04-27 e296cd9
Deploying to gh-pages from @ ropensci-training/saperlipopette@ac4b2ff8af5b11525c0e325011b352474b18c229 🚀
