+++ articles/saperlipopette.md
@@ -27,10 +27,10 @@ case, the corresponding ohshitgit entry.
 library("saperlipopette")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmphxsD2k/file1ddb4cea135d/one-small-change
+#> ℹ Follow along in /tmp/RtmphQje1e/file1cff1ca02a8f/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/RtmphxsD2k/file1ddb4cea135d/one-small-change
+#> /tmp/RtmphQje1e/file1cff1ca02a8f/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
@@ -85,7 +85,7 @@ better with R Markdown that we use for building documentation.
 ``` r
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmphxsD2k/file1ddb48da1a4c/one-small-change
+#> ℹ Follow along in /tmp/RtmphQje1e/file1cff57893bd3/one-small-change
 gert::git_log(repo = path)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -94,7 +94,7 @@ gert::git_log(repo = path)
 #> 2 e227ecc55e421f70b6e30602e6a2ee… Jane … 2023-12-15 15:25:00     2 FALSE "First…
 parent_path2 <- withr::local_tempdir()
 path2 <- exo_one_small_change(parent_path2)
-#> ℹ Follow along in /tmp/RtmphxsD2k/file1ddb65e97adc/one-small-change
+#> ℹ Follow along in /tmp/RtmphQje1e/file1cffdc59b9a/one-small-change
 gert::git_log(repo = path2)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -113,10 +113,10 @@ library("saperlipopette")
 withr::local_language("fr")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmphxsD2k/file1ddb600481f9/one-small-change
+#> ℹ Follow along in /tmp/RtmphQje1e/file1cff6b30b523/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/RtmphxsD2k/file1ddb600481f9/one-small-change
+#> /tmp/RtmphQje1e/file1cff6b30b523/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
