### 2026-04-27 20b10cb
docs: explain the withr path
