+++ README.Rmd
@@ -64,6 +64,8 @@ You can say "Saperlipopette, Git!".
 ## Example
 
 Running the code below will create a new folder containing the exercise.
+`parent_path` is a temporary folder created by the withr package, that will be deleted when closing the current R session.
+
 If in RStudio or Positron, the code will even open a new R session. 
 If not, the user will need to launch an R session,
 where messages would indicate them what to do, 
