+++ README.md
@@ -63,6 +63,9 @@ say “Saperlipopette, Git!”.
 ## Example
 
 Running the code below will create a new folder containing the exercise.
+`parent_path` is a temporary folder created by the withr package, that
+will be deleted when closing the current R session.
+
 If in RStudio or Positron, the code will even open a new R session. If
 not, the user will need to launch an R session, where messages would
 indicate them what to do, and which URL to follow, to find, in this
@@ -72,10 +75,10 @@ case, the corresponding ohshitgit entry.
 library("saperlipopette")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf7ad80e46/one-small-change
+#> ℹ Follow along in /tmp/Rtmp3fVeqS/fileee183ff154a9/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/RtmpSfbavO/filee7bf7ad80e46/one-small-change
+#> /tmp/Rtmp3fVeqS/fileee183ff154a9/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
@@ -129,7 +132,7 @@ building documentation.
 ``` r
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf7d1ccd7/one-small-change
+#> ℹ Follow along in /tmp/Rtmp3fVeqS/fileee187fbd5a93/one-small-change
 gert::git_log(repo = path)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -138,7 +141,7 @@ gert::git_log(repo = path)
 #> 2 e227ecc55e421f70b6e30602e6a2ee… Jane … 2023-12-15 16:25:00     2 FALSE "First…
 parent_path2 <- withr::local_tempdir()
 path2 <- exo_one_small_change(parent_path2)
-#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf79c2b66/one-small-change
+#> ℹ Follow along in /tmp/Rtmp3fVeqS/fileee185d0d874e/one-small-change
 gert::git_log(repo = path2)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -157,10 +160,10 @@ library("saperlipopette")
 withr::local_language("fr")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmpSfbavO/filee7bf36801ecc/one-small-change
+#> ℹ Follow along in /tmp/Rtmp3fVeqS/fileee184cf6259b/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/RtmpSfbavO/filee7bf36801ecc/one-small-change
+#> /tmp/Rtmp3fVeqS/fileee184cf6259b/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
