### 2026-04-27 cb0bcf6
Deploying to gh-pages from @ ropensci-training/saperlipopette@33fa5ff9dd177ae34fc7e8f07cc0609b162d0c85 🚀
