+++ articles/saperlipopette.md
@@ -18,6 +18,9 @@ exclamation](https://en.wiktionary.org/wiki/saperlipopette). You can say
 ## Example
 
 Running the code below will create a new folder containing the exercise.
+`parent_path` is a temporary folder created by the withr package, that
+will be deleted when closing the current R session.
+
 If in RStudio or Positron, the code will even open a new R session. If
 not, the user will need to launch an R session, where messages would
 indicate them what to do, and which URL to follow, to find, in this
@@ -27,10 +30,10 @@ case, the corresponding ohshitgit entry.
 library("saperlipopette")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmphQje1e/file1cff1ca02a8f/one-small-change
+#> ℹ Follow along in /tmp/Rtmpi45mqf/file1d321e5269b7/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/RtmphQje1e/file1cff1ca02a8f/one-small-change
+#> /tmp/Rtmpi45mqf/file1d321e5269b7/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
@@ -85,7 +88,7 @@ better with R Markdown that we use for building documentation.
 ``` r
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmphQje1e/file1cff57893bd3/one-small-change
+#> ℹ Follow along in /tmp/Rtmpi45mqf/file1d325ab9b95b/one-small-change
 gert::git_log(repo = path)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -94,7 +97,7 @@ gert::git_log(repo = path)
 #> 2 e227ecc55e421f70b6e30602e6a2ee… Jane … 2023-12-15 15:25:00     2 FALSE "First…
 parent_path2 <- withr::local_tempdir()
 path2 <- exo_one_small_change(parent_path2)
-#> ℹ Follow along in /tmp/RtmphQje1e/file1cffdc59b9a/one-small-change
+#> ℹ Follow along in /tmp/Rtmpi45mqf/file1d326d189c1a/one-small-change
 gert::git_log(repo = path2)
 #> # A tibble: 2 × 6
 #>   commit                          author time                files merge message
@@ -113,10 +116,10 @@ library("saperlipopette")
 withr::local_language("fr")
 parent_path <- withr::local_tempdir()
 path <- exo_one_small_change(parent_path)
-#> ℹ Follow along in /tmp/RtmphQje1e/file1cff6b30b523/one-small-change
+#> ℹ Follow along in /tmp/Rtmpi45mqf/file1d3238d50da3/one-small-change
 # what's in path
 fs::dir_tree(path)
-#> /tmp/RtmphQje1e/file1cff6b30b523/one-small-change
+#> /tmp/Rtmpi45mqf/file1d3238d50da3/one-small-change
 #> ├── R
 #> └── bla
 # with Git in a command line: git log
