### 2026-04-27 3aebdb4
Deploying to gh-pages from @ ropensci-training/saperlipopette@20b10cb0cf3ee163fd8b4372917d3a022ed7af13 🚀
