### 2026-04-27 #754 @maelle
Thanks a lot @jcoliver and @mine-cetinkaya-rundel for your thorough reviews of saperlipopette!

# Response to @jcoliver

> This is a useful package when paired with published lessons; however, it took
me (as a naive user) a while to get to that important point. (It was not made
clear to me until the "Teaching with saperlipopette" section of the second
vignette). There is a brief mention of this early on in the README; that is,
lines 24-25 of README.Rmd hint at the use, but additional information would be
useful.

I [added](https://github.com/ropensci-training/saperlipopette/pull/50/changes/aa279a581babe47cfe8e73b1ce513bb3f64d013c) those lines at the top of the README

```md
In practice here, in the exercise's folder, the user would
- change a file,
- then Git add it through the terminal or the interface of their choice,
- then run `git commit --amend --no-edit` through the terminal or the interface of their choice.

The user would examine the [Git history](https://git-scm.com/book/en/v2/Git-Basics-Viewing-the-Commit-History) before and after this, either through the terminal (`git log`) or the interface of their choice.
```

> In addition to this clarification, I think an expanded vignette with use cases
is warranted. To be extremely pedantic, I see this point in the Packaging
Guide (1.7.1 point #4):

Regarding a vignette with use cases, what doesn't the article _"saperlipopette, a risk-free playground for learning more Git"_
cover that could be useful?
That article even links to workshops that have been taught using the package.
Do you mean I should for instance add the plan of a lesson to the main vignette?
Also note that learners can also use the package on their own, running the exercises of their choice.

>  I think a note about the coming behavior should be added before
the code block, so the user has a better sense of what is about to happen. This
could be done by moving lines 42-44 before the code block (and revising the
wording to indicate the things that happen automatically).

I changed the wording around the example in the README and vignette.
[Commit 1](https://github.com/ropensci-training/saperlipopette/pull/50/changes/cbf9099e3c9a67b4be028e97b0075f68df7aae8a),
[commit 2](https://github.com/ropensci-training/saperlipopette/pull/50/changes/e71ed48471ef0cbbfe26f23ceecc39e5987cfbf3)

> Related, all functions that start a new R (and RStudio) session should
explicitly say this behavior (starting a new R session) is expected (probably
best to be done in Details sections of function documentation). Yes, I
recognize this would be repetitive, but I think it is useful for new users.

I [added](https://github.com/ropensci-training/saperlipopette/pull/50/changes/82a96564b711d9b4c36c1525184a72d59d724dde) a workflow session to all manual pages.

> create_all_exercises. I don't know when one would use this function. A use case
(not just an example) would be helpful.

This came from [early feedback](https://github.com/ropensci-training/saperlipopette/issues/9) on the package.
I [added](https://github.com/ropensci-training/saperlipopette/pull/50/changes/ae875ae57b91aee529215f37a3c849a55028cc08) a sentence at the top of the manual page.

> The following files include functions that lack documentation for use (even
though most are for internal use only):

I added comments or links for all of them:

- [roxygen2.R](https://github.com/ropensci-training/saperlipopette/pull/50/changes/f4c7be6c539ef5693510d2374d7fb09d82f4f7f9) (opportunity to also simplify the infrastructure)
- [utils-fs.R](https://github.com/ropensci-training/saperlipopette/pull/50/changes/6cba620fbe86a4d3be89ec0821225c557ca183f5)
- [utils-git.R](https://github.com/ropensci-training/saperlipopette/pull/50/changes/5f56f85a27455734dc5e9aa9c04a2a379d773f7c) (opportunity to use better names)
- [utils-potools.R](https://github.com/ropensci-training/saperlipopette/pull/50/changes/f405a794066f3264ef2f077eb8c714ca063c06ad)
- [utils-usethis.R](https://github.com/ropensci-training/saperlipopette/pull/50/changes/fb2f831961bb3f350d1870e0f352383f446fc289)

> Not all files are named the same as their functions.

[Done](https://github.com/ropensci-training/saperlipopette/pull/50/changes/b21a10b29164adb972502c228b34127565f77adf) with the help of Claude Sonnet!

> Consider using return().

I prefer not to, following the [tidyverse style guide](https://style.tidyverse.org/functions.html#return).

> In the "Follow along..." messages, consider removing the trailing exclamation
point (!). Users may think this needs to be part of the path.

Good (exclamation) point, [done](https://github.com/ropensci-training/saperlipopette/pull/50/changes/4661c0a676a992399dbcdb26431cb2e0d84e160a)!

> Me, acting very naively, identified an opportunity for some defensive
programming: If I try creating all exercises after running any of the other,
individual exercise functions (exo_*), it throws an error (appropriately so,
perhaps). For example, I first ran exo_one_small_change(parent_path) as in
the example, then ran create_all_exercises(parent_path)

I [added](https://github.com/ropensci-training/saperlipopette/pull/50/changes/17b93a257adac99552594c9818354fb0308f44f9) a caveat to the manual page.

> Is current release (7.3.3) not sufficient?

Actually yes it is, now [changed](https://github.com/ropensci-training/saperlipopette/pull/50/changes/b62a1453558523de2053d70176b12d2ffa215287)!

> Also, the run of devtools::check() failed during the installation step when
trying to install the dependency parsedate. I went ahead and manually installed
parsedate, and the check proceeded without additional problems. It is entirely
possible I forgot to do something before running devtools::check()

parsedate is listed as a dependency in `DESCRIPTION`, hopefully this was a glitch?
I haven't heard from other users about this problem in particular.

> Perhaps this version dependency should be added to the DESCRIPTION? e.g.

[Done](https://github.com/ropensci-training/saperlipopette/pull/50/changes/82cf2bafbb108764e873fcff6a2670e695cbcd1c)

# Response to @mine-cetinkaya-rundel

> The justification for the package name is nice, though I might suggest also noting that the Wikipedia link includes pronunciation.

[Mention added](https://github.com/ropensci-training/saperlipopette/commit/ac4b2ff8af5b11525c0e325011b352474b18c229).
For the record, I slightly regret the name and gave a simpler name to saperlipopette's CLI sibling, [zut](https://masalmon.eu/2026/02/27/zut-saperlipopette-claude/).

> The working example demonstrating the exercise workflow... works! I tried in Positron and had no issues. However, it's likely that a user who needs Git exercises isn't familiar with the withr package. I recommend explaining the code in the first few lines of the example before the example in the README, and also noting that this will start a new Positron session.

Explanation added for the [withr path](https://github.com/ropensci-training/saperlipopette/commit/33fa5ff9dd177ae34fc7e8f07cc0609b162d0c85)
and the [new session](https://github.com/ropensci-training/saperlipopette/commit/59bb6b9c638f25998aeda14efb58ed19f98df588) (see also my response to the other reviewer).

> Main vignette (vignettes/saperlipopette.qmd) provides a comprehensive introduction that is basically the same as the package README, so the same comments apply here.

I made the changes in the two places.

> Additionally, the rendered vignette on the package website has "[31m✖[39m", which I believe is a Quarto issue, but it's distracting nonetheless. Is it possible to suppress them?

Ouch, thanks for reporting!
I believe it's an R-universe problem, and opened an [issue to track it](https://github.com/ropensci-org/rotemplate/issues/129).

> As the other reviewer suggested, expanding documentation for create_all_exercises() would be helpful.

This came from [early feedback](https://github.com/ropensci-training/saperlipopette/issues/9) on the package.
I [added](https://github.com/ropensci-training/saperlipopette/pull/50/changes/ae875ae57b91aee529215f37a3c849a55028cc08) a sentence at the top of the manual page.

> DESCRIPTION specifies RoxygenNote: 7.3.3.9000 (development version), but the released version 7.3.3 is typically installed. Is there a reason the dev version is needed? Otherwise,c I recommend sticking to the released version.

Now [changed](https://github.com/ropensci-training/saperlipopette/pull/50/changes/b62a1453558523de2053d70176b12d2ffa215287)!

# Thank you

Thanks again @jcoliver and @mine-cetinkaya-rundel, your feedback has been most helpful. :pray:
