### 2026-04-27 #754 @robitalec
Thanks @mine-cetinkaya-rundel for your thoughtful review. I really appreciate you highlighting potential knowledge gaps that might prevent a user from fully taking advantage of the exercises provided by saperlipopette.

I agree with both reviewers that somehow improving the accessibility of the teaching information in saperlipopette would be beneficial. Maybe this looks like taking some of the information from the vignette into additional vignettes "Teaching with saperlipopette" and (individual) "Learning with saperlipopette". It might help a user or instructor to have a information e.g. a skeleton lesson plan or learning objectives associated with exercises framed specifically for each use case.
