### 2026-04-27 #754 @robitalec
Thanks @maelle for quickly taking into consideration the reviews from @jcoliver and @mine-cetinkaya-rundel with your changes.

We're ready for reviewer responses, so @jcoliver and @mine-cetinkaya-rundel please use the [Approval Template](https://devguide.ropensci.org/approval2template.html) to communicate if @maelle has sufficiently addressed your comments. More details are available in the Guide for Reviewers [here](https://devguide.ropensci.org/softwarereview_reviewer.html#followupreviewer).

As always, please use this thread for any questions or comments along the way.
