### 2026-05-04 8cc97a8
docs: add two small vignettes
