+++ vignettes/articles/teaching.Rmd
@@ -1,0 +1,47 @@
+---
+title: "Teaching with saperlipopette"
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r setup}
+library(saperlipopette)
+```
+
+The saperlipopette package was created with teaching Git in mind!
+Refer to the article ["saperlipopette, a risk-free playground for learning more Git"](https://docs.ropensci.org/saperlipopette/articles/paper.html) for more context.
+Its references include links to materials of tutorials/workshops using saperlipopette.
+
+In practice here's how we teach with saperlipopette:
+
+## Preparation instructions
+
+We prepare installation instructions before the workshop, including Git and saperlipopette. In theory, saperlipopette is used to teach Git to people who already use it so they probably have some sort of Git set up.
+
+## Curriculum
+
+We choose the focus of the session, for instance fixing common mistakes, and pick relevant exercises.
+
+In particular, we've taught
+
+- [how to improve one's Git history](https://masalmon.eu/talks/2025-11-24-git-history/);
+- [how to fix common Git mistakes](https://masalmon.eu/talks/2026-03-19-oops-git/)
+
+but you could focus on any topic, small or large.
+You can re-use (even better if you acknowledge our work) our materials, mix and match them.
+
+We order and group those exercises. For instance fixing the latest commit's message or content are two very similar exercises.
+
+## During the workshop
+
+During the workshop, a bit of context is explained in each part but the focus is on demonstrating each exercise or each pair/group of exercises then letting participants solve the exercises on their own. 
+We strongly believe hands-on sessions are key to learning.
+
+## After the workshop
+
+Or before... please share with us any comment, bug report or feature request on the exercises!
