### 2026-05-04 2ea42ce
Apply suggestions from code review
Co-authored-by: Yanina Bellini Saibene <yabellini@gmail.com>
Co-authored-by: Maëlle Salmon <maelle.salmon@yahoo.se>
