### 2026-05-04 ee78883
docs: add two small vignettes (#51)
* docs: add two small vignettes

* Apply suggestions from code review

Co-authored-by: Yanina Bellini Saibene <yabellini@gmail.com>
Co-authored-by: Maëlle Salmon <maelle.salmon@yahoo.se>

---------

Co-authored-by: Yanina Bellini Saibene <yabellini@gmail.com>
