### 2026-05-04 #754 @maelle
Even before getting the answer from @jcoliver, we've added two vignettes to the package following the structure suggested by @robitalec https://github.com/ropensci-training/saperlipopette/pull/51 😸
