### 2026-05-04 2b60247
Deploying to gh-pages from @ ropensci-training/saperlipopette@ee788837c059e560642f7d538c4b173eb03f117f 🚀
