+++ articles/learning.md
@@ -1,0 +1,31 @@
+# Learning with saperlipopette
+
+You can use the saperlipopette R package to practice Git in a safe
+environment!
+
+There are different ways to go about this:
+
+- You could try to do all the exercises. To create all the exercises in
+  one go, see the
+  [`create_all_exercises()`](https://docs.ropensci.org/saperlipopette/reference/create_all_exercises.md)
+  function. Maybe one day we’ll have enough exercises for an Advent
+  calendar of Git!
+
+- Say you hear about Git worktrees in a [blog
+  post](https://masalmon.eu/2026/04/23/git-worktree-again/) or
+  [tool](https://docs.ropensci.org/saperlipopette/articles/github.com/DavisVaughan/cross/).
+  You could decide to practice using the relevant exercise after
+  searching for “worktree” in the pkgdown website:
+  [`exo_worktree()`](https://docs.ropensci.org/saperlipopette/reference/exo_worktree.html).
+
+- You need to do an interactive rebase but are a bit unsure about it.
+  You could first refresh your knowledge using
+  [`exo_rebase_i()`](https://docs.ropensci.org/saperlipopette/reference/exo_rebase_i.html)
+  before using the skill on a real repository!
+
+For each exercise, you could do them using the `tip()` function the
+first time to get advice on how to solve it, and refraining from using
+it the second time, to check your understanding.
+
+Please share with us any comment, bug report or feature request on the
+exercises!

[…]
