### 2026-05-04 #754 @jcoliver
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 1 (total: 5)

The majority of my comments have been addressed. In regards to the vignette I mentioned, I had in mind something that a user (teacher or self-directed learner) could guide their first use of the package. The article "saperlipopette, a risk-free playground for learning more Git" provides high-level information and justification, but is not something a new user to saperlipopette could really use to get started. I think the introductory vignettes from [dplyr](https://cran.r-project.org/web/packages/dplyr/vignettes/dplyr.html) and [dbplyr](https://cran.r-project.org/web/packages/dbplyr/vignettes/dbplyr.html) do a better job showing what I am trying to explain. I think expanding the "Get started" vignette, borrowing from the material on Zenodo could go a long way to communicating worked examples of how to use the package.
