### 2026-05-05 #754 @mine-cetinkaya-rundel
## Reviewer Response

#### Final approval (post-review)

Estimated hours spent reviewing: 0.5 (total: 3.5)

My comments have been addressed. I think improvements that have made based on both sets of review comments make the package much easier to get started with for educators.
