### 2026-05-06 #754 @robitalec
Thank you to @maelle for the submission and a huge thank you to @jcoliver and @mine-cetinkaya-rundel for your reviews of saperlipopette. I really appreciate your thoughtful comments and contributions to this package. Your experience and perspectives clearly highlighted places where @maelle could improve the package, and I think it's more accessible to potential users as a result.
